[🚚Rename] : 폴더 이름 수정 1. 폴더 이름 수정
