+++ src/pages/SelectLoginPage/SelectLoginPage.jsx
@@ -1,7 +1,7 @@
 import { Wish } from '@/components/atoms';
 import Divider from '@/components/atoms/Divider/Divider';
 import LoginButton from '@/components/atoms/LoginButton/LoginButton';
-import SignUpLink from '@/components/atoms/SignUpLink/SignUpLink';
+import SignUpLink from '@/components/atoms/SignUpZone/SignUpZone';
 import SocialLoginButton from '@/components/atoms/SocialLoginButton/SocialLoginButton';
 import pb from '@/utils/pocketbase';
 import { useState } from 'react';
